An eighth table, and a tuple that dropped a field: the mechanism was neither
Asked whether the mechanism was general and minimal, it was not, and the
first defect was the one the previous commit claimed to have removed.

**`ERROR_MODES` was an eighth dict keyed by workload name.** `Workload` had
eleven fields and not the one the diagnosis needs, so a sixth workload could
still be half-registered -- failing at `assess()`, after a paid run. Exactly
the `_INDEXERS` shape.

It could not be otherwise: `audit_modes.py` imported its scoring primitives
*from* `audit_phase0.py`, so the registry in phase0 could not hold the mode
functions. A circular import was forcing the table. Inverting the dependency
fixes it at the shape rather than at the instance:

* `scripts/audit_workloads.py` -- what a workload **is**: how it is scored,
  how a row becomes a task, how its errors are named, and the registry.
* `scripts/audit_phase0.py` -- the Phase 0 *experiment*, 1177 -> 731 lines,
  which is what it should always have been.

One direction, and `modes` is now a field like the rest.

**The mode functions took a positional `(question, reference, extra)`
tuple.** `task_index` already returns real `Task` objects; the tuple threw
one away to rebuild a worse version of it -- dropping `meta["tests"]`
entirely, so nothing downstream of a mode function could re-run an MBPP
candidate, and giving slot `[2]` a different meaning per workload. It
existed so I would not have to touch `ctx[0]`/`ctx[1]` call sites. That is
debt, not design. A context is the task now.

**A third dict named `WORKLOADS`** in `audit_judge_repair.py`, meaning
something else entirely (name -> records path), is `RECORDS`.

Line count is flat, which is the honest number: this is a move, not a
deletion. The win is that `audit_phase0.py` shed 446 lines of things that
were never about Phase 0, and that a workload is now defined in exactly one
place with no way to register half of it.

Checked: `verifier_diagnosis_f55dec40cec559f7.md` regenerates byte-identical
across the whole restructure, all five workloads dry-run end to end, and
MBPP still clears the rung-5 gate on sufficiency without `--force`.
